+++ client/src/App.tsx
@@ -1,25 +1,14 @@
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
+import { RouterProvider } from 'react-router-dom';
 
 import { AudioContextProvider } from './contexts/AudioContext';
-import Home from './pages/Home/Home';
-import Podcasts from './pages/Podcasts/Podcasts';
-import Albums from './pages/Albums/Albums';
-import RadioStations from './pages/RadioStations/RadioStations';
-
-const router = createBrowserRouter([
-  {path: "/", element: (<Home/>)},
-  {path: '/podcasts', element: (<Podcasts/>)},
-  {path: '/radioStations', element: (<RadioStations/>)},
-  {path: '/albums', element: (<Albums/>)}
-])
-
-
+import AudioPlayer from './layouts/components/AudioPlayer/AudioPlayer';
+import { router } from './routes/router';
 
 function App() {
- 
   return (
-    <AudioContextProvider>
+    <AudioContextProvider>  
        <RouterProvider router={router} />
+       <AudioPlayer/>
     </AudioContextProvider>
   );
 }
